Request wildcard certs with dns-01 challenges, and various updates.
This replaces the original http-01 storage-account-blob function with
one that uses Azure DNS to complete a dns-01 challenge.

With this change, the dns-01 function has no connection to the storage account
used by the CDN function, so the codebase has been refactored to be two
separate functions.

Other changes:

- Update to .Net Core 3.1, and update Azure Functions deps.

- Use nuget lockfile so that newer deps don't break the build.

- Build in Docker so that dotnet and func don't need to be managed on the host.

  This also automates many of the manual steps in the README.

- Restore ability to use SP, since MSI can't be used for local testing.

  This reverts a206f65dfd89d3eb4c84b3a65eb49590b0a8ce37 to some extent.

Fixes #1