Use ARI when available over hard-coded 30-day renewal period.
We intentionally do not use the ARI ID of the existing cert as
the `replaces` field of a new order, because LE at least rejects such an order
as "unauthorized" if the existing cert was not issued by the same account key.
Since we do not track which account key was used to issue the existing cert,
the only alternative would be to retry the request without the `replaces` field.
The benefit of using `replaces` does not seem to justify this complexity.

Fixes #9